+++ lib/zabbix-ruby-client/plugins/rabbitmq.rb
@@ -1,9 +1,7 @@
-# read the remote data from http://mq-dev.faria.co/
+# read the remote data with http://www.rabbitmq.com/management-cli.html
 require "zabbix-ruby-client/logger"
 require "zabbix-ruby-client/plugin_base"
 require "json"
-require "net/http"
-require "curb"
 
 module ZabbixRubyClient
   module Plugins
@@ -13,86 +11,44 @@ module ZabbixRubyClient
 
       def collect(*args)
         host = args[0]
-        url = "http://host/api/overview"
-        username = "user"
-        userpwd = "pass"
-        info = get_info(url,username,userpwd)
+        rabbitmqadmin = args[1]
+        login = args[2]
+        pass = args[3]
+        info = get_info(rabbitmqadmin, login, pass)
         back = []
         if info
           time = Time.now.to_i
-          back << "#{host} rabbitmq[received] #{time} #{info['publish']}"
-          back << "#{host} rabbitmq[sent] #{time} #{info['ack']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['deliver_get']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['redeliver']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['deliver']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['messages']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['messages_ready']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['messages_unacknowledged']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['consumers']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['queues']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['exchanges']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['connections']}"
-          back << "#{host} rabbitmq[count] #{time} #{info['channels']}"
+          back << "#{host} rabbitmq.erlang.version #{time} #{info['erlang_version']}"
+          back << "#{host} rabbitmq.message.ack #{time} #{info['message_stats']['ack']}"
+          back << "#{host} rabbitmq.message.ack.rate #{time} #{info['message_stats']['ack_details']['rate']}"
+          back << "#{host} rabbitmq.message.deliver #{time} #{info['message_stats']['deliver']}"
+          back << "#{host} rabbitmq.message.deliver.rate #{time} #{info['message_stats']['deliver_details']['rate']}"
+          back << "#{host} rabbitmq.message.deliver_get #{time} #{info['message_stats']['deliver_get']}"
+          back << "#{host} rabbitmq.message.deliver_get.rate #{time} #{info['message_stats']['deliver_get_details']['rate']}"
+          back << "#{host} rabbitmq.message.deliver_get #{time} #{info['message_stats']['deliver_get']}"
+          back << "#{host} rabbitmq.message.deliver_get.rate #{time} #{info['message_stats']['deliver_get_details']['rate']}"
+          back << "#{host} rabbitmq.message.publish #{time} #{info['message_stats']['publish']}"
+          back << "#{host} rabbitmq.message.publish.rate #{time} #{info['message_stats']['publish_details']['rate']}"
+          back << "#{host} rabbitmq.message.redeliver #{time} #{info['message_stats']['redeliver']}"
+          back << "#{host} rabbitmq.message.redeliver.rate #{time} #{info['message_stats']['redeliver_details']['rate']}"
+          back << "#{host} rabbitmq.queue.total.messages #{time} #{info['queue_totals']['messages']}"
+          back << "#{host} rabbitmq.queue.total.messages_ready #{time} #{info['queue_totals']['messages_ready']}"
+          back << "#{host} rabbitmq.queue.total.messages_unacknowledged #{time} #{info['queue_totals']['messages_unacknowledged']}"
+          back << "#{host} rabbitmq.total.channels #{time} #{info['object_totals']['channels']}"
+          back << "#{host} rabbitmq.total.connections #{time} #{info['object_totals']['connections']}"
+          back << "#{host} rabbitmq.total.consumers #{time} #{info['object_totals']['consumers']}"
+          back << "#{host} rabbitmq.total.exchanges #{time} #{info['object_totals']['exchanges']}"
+          back << "#{host} rabbitmq.total.listeners #{time} #{info['listeners'].count}"
+          back << "#{host} rabbitmq.total.queues #{time} #{info['object_totals']['queues']}"
+          back << "#{host} rabbitmq.version #{time} #{info['rabbitmq_version']}"
         end
         back
       end
 
     private
 
-      def get_info(url,username,userpwd)
-
-        curl = Curl::Easy.new(url)
-        curl.username = username
-        curl.password = userpwd
-        curl.perform
-        data_obj = JSON.parse(curl.body_str)
-
-        ret = {}
-        data_obj["message_stats"].each do |dshow|
-          ret[dshow[0]]=dshow[1]
-        end
-        data_obj["queue_totals"].each do |qshow|
-          ret[qshow[0]]=qshow[1]
-        end
-        data_obj["object_totals"].each do |oshow|
-          ret[oshow[0]]=oshow[1]
-        end
-
-        # puts ret["publish"]
-        # puts ret["ack"]
-        # puts ret["deliver_get"]
-        # puts ret["redeliver"]
-        # puts ret["deliver"]
-        # puts ret["messages"]
-        # puts ret["messages_ready"]
-        # puts ret["messages_unacknowledged"]
-        # puts ret["consumers"]
-        # puts ret["queues"]
-        # puts ret["exchanges"]
-        # puts ret["connections"]
-        # puts ret["channels"]
-
-        ret
-#        info = getlines(data_obj, "[^,]*,[^,]*,.* [0-9]{4}")
-#        if info
-#          received = 0
-#          sent = 0
-#          count = 0
-#          info.each do |line|
-#            if line[0] =~ /^[a-z]/
-#              back = line.split(',')
-#              received += back[2].to_i
-#              sent += back[3].to_i
-#              count += 1
-            # end
-          # end
-          # ret["received"] = received
-          # ret["sent"] = sent
-          # ret["count"] = count
-          # ret
-        # else
-          # false
-        # end
+      def get_info(rabbitmqadmin, login, pass)
+        JSON.parse(`#{rabbitmqadmin} -u #{login} -p #{pass} -f raw_json show overview`)
       end
 
     end